requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
httpx>=0.25.2,<0.26.0
PyGithub>=2.0.0
mistralai>=0.1.0
python-multipart>=0.0.6
cryptography>=41.0.0